Enable router and create new note component
